requirements.txt
Flask==2.0.1
python-telegram-bot==13.15
pymongo==4.0.1
python-dotenv==0.19.0